applying review suggestions
